requirements.txt
Django>=2.1.3
gunicorn>=19.9.0
psycopg2-binary
dj-database-url==0.4.2
whitenoise>=4.1
#Pillow>=5.3.0
Markdown>=2.6.8
django-pagedown>=0.1.3
django-taggit
django-vote
bleach>=2.0.0
python-dotenv==0.7.1
click>=7.0
boto3>=1.9.57
botocore>=1.12.57
django-storages>=1.5.2
mixpanel>=4.3.2
zappa>=0.51.0
django-push
feedparser==5.2.1
nltk
numpy
networkx
beautifulsoup4

#API
djangorestframework
djangorestframework-jwt
django-cors-headers>=3.2.1
#git+https://github.com/underlost/Zappa.git
